Implemented new design
