+++ packages/esm-billing-app/translations/en.json
@@ -2,11 +2,11 @@
   "actions": "Actions",
   "active": "Active",
   "addBill": "Add bill item(s)",
-  "addCommodity": "Add commodity",
+  "addCommodityChargeItem": "Add charge item",
   "addPaymentOptions": "Add payment option",
-  "addService": "Add service",
+  "addServiceChargeItem": "Add charge item",
   "amount": "Amount",
-  "billableService": "Billable services table",
+  "applyFilter": "Apply filter",
   "billableServicesError": "Billable services error",
   "billAmount": "Bill Amount",
   "billedItems": "Billed Items",
@@ -25,10 +25,9 @@
   "billsList": "Bill list",
   "billTotal": "Bill total",
   "cancel": "Cancel",
-  "chargeableDashboard": "Chargeable Dashboard",
-  "chargeableItems": "Chargeable Items",
-  "chargeableServices": "Chargeable Summary",
-  "chargeableServicesDescription": "Chargeable services and consumable prices",
+  "chargeItem": "Charge items table",
+  "chargeItems": "Charge Items",
+  "chargeItemsDescription": "Charge Items and Services Dashboard",
   "checkFilters": "Check the filters above",
   "claim": "Process claims",
   "claimcode": " Claim Code",
@@ -45,10 +44,9 @@
   "discard": "Discard",
   "discardClaim": "Discard Claim",
   "discount": "Discount",
-  "edit": "Edit",
-  "editBillableService": "Edit Billable Service",
   "editBillForm": "Edit Bill Form",
-  "editCommodity": "Edit Commodity",
+  "editChargeItem": "Edit charge item",
+  "editServiceChargeItem": "Edit Service Charge Item",
   "enterAmount": "Enter amount",
   "enterPolicyNumber": "Enter policy number",
   "enterPrice": "Enter price",
@@ -128,9 +126,10 @@
   "providerMessage": "By clicking Proceed to care, you acknowledge that you have advised the patient to settle the bill.",
   "quantity": "Quantity",
   "referenceNumber": "Ref number",
+  "resetFilters": " Reset filters",
   "saveAndClose": "Save & Close",
   "search": "Search",
-  "searchForBillableService": "Search for billable service",
+  "searchForChargeItem": "Search for charge item",
   "searchForCommodity": "Search for commodity ",
   "searchForConcept": "Search for concept",
   "searchPlaceHolder": "Find your billables here...",
