+++ packages/esm-patient-clinical-view-app/src/peer-calendar/header/peer-calendar-header.scss
@@ -1,7 +1,6 @@
 @use '@carbon/layout';
 @use '@carbon/type';
 @use '@carbon/colors';
-@import '~@openmrs/esm-styleguide/src/vars';
 
 .reportingPeriod {
   display: flex;
@@ -9,7 +8,7 @@
   flex-flow: row wrap;
   gap: layout.$spacing-01;
   width: 100%;
-  padding: layout.$spacing-03 0 0 0;
+  padding: layout.$spacing-03 layout.$spacing-05 0 0;
 }
 
 .reportingPeriodInput {
@@ -72,47 +71,8 @@
   @include type.type-style('label-01');
 }
 
-
 .illustration {
   width: 72px;
   height: 72px;
   fill: var(--brand-03);
-}
-
-.btnSet {
-  display: flex;
-  justify-content: space-between;
-  margin-top: layout.$spacing-05;
-  margin-bottom: layout.$spacing-05;
-  flex-flow: row wrap;
-  align-items: center;
-}
-
-.btnSetRight {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.detailHeaderContainer {
-  margin: layout.$spacing-05;
-  flex-flow: row wrap;
-  justify-content: space-between;
-}
-
-.detailHeaderContent {
-  background-color: colors.$gray-10;
-  width: fit-content;
-}
-
-.detailHeaderContentLoading {
-  background-color: colors.$gray-10;
-  max-width: 70%;
-}
-
-.detailHeaderContentRow {
-  display: flex;
-  flex-flow: row wrap;
-  gap: layout.$spacing-08;
-  align-items: center;
 }
