+++ packages/esm-patient-clinical-view-app/src/peer-calendar/metrics/metrics.scss
@@ -1,0 +1,50 @@
+@use '@carbon/layout';
+@use '@carbon/type';
+@use '@carbon/colors';
+
+.btnSet {
+  display: flex;
+  justify-content: space-between;
+  margin-top: layout.$spacing-05;
+  flex-flow: row wrap;
+  align-items: center;
+}
+
+.btnSetRight {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+
+.detailHeaderContainer {
+  flex-flow: row wrap;
+  justify-content: space-between;
+  padding: layout.$spacing-05;
+}
+
+.metricsContainer {
+  display: flex;
+  flex-flow: row wrap;
+  width: 100%;
+  flex-grow: 1;
+  gap: layout.$spacing-05;
+}
+
+.metricCard {
+  display: flex;
+  flex-direction: column;
+  gap: layout.$spacing-03;
+  flex-grow: 1;
+  border: solid 1px colors.$gray-20;
+  color: colors.$gray-70;
+  min-width: 320px;
+  background-color: colors.$white;
+
+  & > span {
+    font-size: xx-large;
+  }
+
+  & > p {
+    font-size: small;
+  }
+}
