+++ packages/esm-patient-clinical-view-app/translations/en.json
@@ -34,7 +34,7 @@
   "clientEligibility": "Eligible",
   "clinicalEncounter": "Clinical Encounter",
   "clinicalViews": "Clinical views",
-  "completed": "Completed:",
+  "completed": "Completed",
   "contact": "Contact",
   "contactCreated": "Contact created",
   "contacted": "Contacted",
@@ -66,6 +66,7 @@
   "encountersDescription": "List of encounters during the current visit",
   "encountersList": "Encounters list",
   "encounterType": "Encounter Type",
+  "enterForm": "Enter form",
   "error": "Error",
   "errorFetchingbedInformation": "Error fetching bed information",
   "facility": "Facility",
@@ -96,7 +97,6 @@
   "inPatientVisitMessage": "in-patient encounter found for current  {{visitType}} visit",
   "ipvOutcome": "IPV Outcome",
   "labourAndDelivery": "Labour and Delivery",
-  "launch": "Launch outreact form",
   "listingDate": "Listing date",
   "livingWithClient": "Living with client",
   "managePeers": "Manage Peers",
@@ -160,10 +160,11 @@
   "testStrategy": "Strategy",
   "therapiesPrescribed": "Therapies Prescribed",
   "timeRecorded": "Time Recorded",
-  "total": "Total:",
+  "total": "Total",
   "tracingNumber": "Tracing No.",
   "tracingType": "Tracing Type",
   "typeOfSurgery": "Type of Surgery",
+  "updateForm": "Mordify form",
   "uterusCondition": "PPH Condition of uterus",
   "viewBeds": "View beds",
   "viewEncounter": "View",
